Implemented Baseclass for education feature
